Bifunctor organized into internal modules
